refactor: move types and api code into component folder
