+++ src/components/Users/types.ts
@@ -1,0 +1,15 @@
+export type Group = {
+    createdAt: string,
+    hostname: string,
+    name: string,
+    updatedAt: string,
+}
+
+export type User = {
+    adminGroups: Array<Group>,
+    createdAt: string,
+    email: string,
+    groups: Array<Group>,
+    id: number,
+    updatedAt: string
+}
